Added inputs validation

+++ universal-login-react/src/core/services/DeleteAccountService.ts
@@ -2,19 +2,40 @@ import {utils} from 'ethers';
 import {transactionDetails} from '../../core/constants/TransactionDetails';
 import {DeployedWallet} from '@universal-login/sdk';
 
-const validateUsername = (usernameInput: string, usernameLogged: string): boolean =>
-  usernameInput === usernameLogged;
-
-export const validateVerifyField = (verifyFieldInput: string): boolean => {
-  const verifyFieldText = 'DELETE MY ACCOUNT';
-  return verifyFieldInput === verifyFieldText;
+type ErrorsType = {
+  usernameError: boolean;
+  verifyFieldError: boolean;
 };
 
-export const deleteAccount = async (deployedWallet: DeployedWallet, onDeleteAccountClick: () => void, username: string, verifyFieldText: string) => {
-  validateUsername(username, deployedWallet.name);
-  validateVerifyField(verifyFieldText);
-  const publicKey = utils.computeAddress(deployedWallet.privateKey);
-  await deployedWallet.removeKey(publicKey, transactionDetails);
-  deployedWallet.sdk.stop();
-  onDeleteAccountClick();
+type InputsType = {
+  username: string,
+  verifyField: string
+};
+
+const validateUsername = (usernameFromInput: string, usernameLogged: string): boolean =>
+  usernameFromInput === usernameLogged;
+
+export const validateVerifyField = (verifyFieldFromInput: string): boolean => {
+  const verifyFieldText = 'DELETE MY ACCOUNT';
+  return verifyFieldFromInput === verifyFieldText;
+};
+
+const checkInputsAgainstError = (usernameLogged: string, inputs: InputsType) => ({
+  usernameError: !validateUsername(inputs.username, usernameLogged),
+  verifyFieldError: !validateVerifyField(inputs.verifyField)
+});
+
+export const getInputClassName = (inputError: boolean) => inputError ? 'delete-account-input-error' : 'delete-account-input';
+
+const doesAnyErrorExists = (errors: ErrorsType) => errors.usernameError || errors.verifyFieldError;
+
+export const deleteAccount = async (deployedWallet: DeployedWallet, inputs: InputsType, setErrors: (errors: ErrorsType) => void, onDeleteAccountClick: () => void) => {
+  const errors = checkInputsAgainstError(deployedWallet.name, inputs);
+  setErrors(errors);
+  if (!doesAnyErrorExists(errors)) {
+    const publicKey = utils.computeAddress(deployedWallet.privateKey);
+    await deployedWallet.removeKey(publicKey, transactionDetails);
+    deployedWallet.sdk.stop();
+    onDeleteAccountClick();
+  }
 };
